# x86-64 host optimizations