--- a/blink/alui.c
+++ b/blink/alui.c
@@ -31,9 +31,7 @@ static void AluiRo(P, const aluop_f ops[4], const aluop_f fast[4]) {
     STATISTIC(++alu_ops);
     switch (GetNeededFlags(m, m->ip, CF | ZF | SF | OF | AF | PF)) {
       case 0:
-      case CF:
-      case ZF:
-      case CF | ZF:
+      CASE_ALU_FAST:
         STATISTIC(++alu_simplified);
         Jitter(A,
                "B"      // res0 = GetRegOrMem(RexbRm)
@@ -76,9 +74,7 @@ static void AluiUnlocked(P, u8 *p, aluop_f op) {
                "r0D",  // PutRegOrMem(RexbRm, res0)
                kJustAlu[ModrmReg(rde)]);
         break;
-      case CF:
-      case ZF:
-      case CF | ZF:
+      CASE_ALU_FAST:
         STATISTIC(++alu_simplified);
         Jitter(A,
                "q"     // arg0 = sav0 (machine)