On my local PC, this patch gives a small (1.4%&mdash;1.8%) but consistent improvement in the running time of the `third_party/cosmo/2/test_suite_mpi.com` test program.  E.g.
```
PASSED (323 / 323 tests (1 skipped))
RL: took 20,104,433µs wall time
RL: ballooned to 5,740kb in size
RL: needed 20,096,321µs cpu (0% kernel)
RL: caused 887 page faults (99% memcpy)
RL: 885 context switches (0% consensual)
RL: performed 0 reads and 8 write i/o operations
```
vs. (without the patch)
```
PASSED (323 / 323 tests (1 skipped))
RL: took 20,407,807µs wall time
RL: ballooned to 5,640kb in size
RL: needed 20,399,361µs cpu (0% kernel)
RL: caused 889 page faults (99% memcpy)
RL: 661 context switches (0% consensual)
RL: performed 0 reads and 8 write i/o operations
```