@tkchia 2023-05-18
With https://github.com/jart/blink/pull/137/commits/98079d733a0f76d5e9b2fb545df711a8512811f3 , a further whopping 7.4% improvement:
```
PASSED (323 / 323 tests (1 skipped))
RL: took 18,715,678µs wall time
RL: ballooned to 5,784kb in size
RL: needed 18,708,731µs cpu (0% kernel)
RL: caused 841 page faults (99% memcpy)
RL: 577 context switches (1% consensual)
RL: performed 0 reads and 8 write i/o operations
```