@tkchia 2023-05-19
Simply disabling the use of Intel control-flow protection instructions (https://github.com/jart/blink/pull/137/commits/309735b8fdb162908efe0a78fc592030a676d788) gives instead an improvement of about 6.9%.  E.g.
```
PASSED (323 / 323 tests (1 skipped))
RL: took 18,778,782µs wall time
RL: ballooned to 5,696kb in size
RL: needed 18,699,182µs cpu (0% kernel)
RL: caused 843 page faults (98% memcpy)
RL: 707 context switches (1% consensual)
RL: performed 1,176 reads and 8 write i/o operations
```

The Intel CET instructions (e.g. `endbr64`) promise better protection against code execution cyber-attacks &mdash; but the protection only occurs if both the OS and CPU know about such instructions.  It seems Linux kernel support for Intel CET is not quite in the main stream yet.