@tkchia 2023-05-20
A further slight improvement of about 1% in the running time (https://github.com/jart/blink/pull/137/commits/b36c201ec239f15b745be286db4ef3e7f550ebd6):
```
PASSED (323 / 323 tests (1 skipped))
RL: took 18,544,090µs wall time
RL: ballooned to 5,572kb in size
RL: needed 18,536,959µs cpu (0% kernel)
RL: caused 838 page faults (99% memcpy)
RL: 539 context switches (1% consensual)
RL: performed 0 reads and 8 write i/o operations
```

I also experimented with using `lahf` instead of `pushfq` + `pop` _reg64_, but this does not appear to help much.