@tkchia 2023-05-21
```
PASSED (323 / 323 tests (1 skipped))
RL: took 18,403,306µs wall time
RL: ballooned to 5,680kb in size
RL: needed 18,354,083µs cpu (0% kernel)
RL: caused 843 page faults (98% memcpy)
RL: 564 context switches (2% consensual)
RL: performed 1,176 reads and 8 write i/o operations
```