@ghaerr 2023-05-25
Hello @tkchia,

Nice job recognizing and getting rid of the `endbr64` instructions and the resulting speedup. Your `ALU_FAST` macros show some deep knowledge of gcc's `asm` directive too - looks like your probably spent a bit of time getting those working :)

Wha's the issue with musl cross-make, does `epoll-wait2` not work on musl?

Thank you!