@tkchia 2023-05-25
Hello @ghaerr,

> looks like your probably spent a bit of time getting those working :)

Yeah, I got bitten by a missing `&` constraint modifier a while back in a different project. :neutral_face: 

> What's the issue with musl cross-make, is `epoll-wait2` not implemented on musl?

Yes, the function is not declared in the pre-built `musl` cross toolchains.

A more "proper" way to fix this would probably be to do a separate `./configure` pass for each of the `$(ARCHITECTURES)` to cross-compile to.  But then again, this seems overkill, since the cross toolchains are mainly used for testing at the moment.

Thank you!